requirements.txt
tensorflow~=2.1.0
numpy~=1.16.5
matplotlib~=3.1.1
scikit-learn~=0.23.2
pandas~=1.1.4
tqdm~=4.50.2
umap-learn~=0.4.2